Problem: when "hard re-booted" (e.g. power loss), dietpi suffers a loss of it's configuration files dietpi.txt, config.txt, boot.ini
When DietPi is restartet with a "power cut" (meaning unplug the power [usb] cord and re-plug it again), DietPi copies at boottime the empty config files dietpi.txt, config.txt, boot.ini into the ramdisk, because these hasn't been rewritten on shutdown before. So in the /boot partition and in the ramdisk there a now simple text files with the text "This file has been moved to DietPi-Ramdisk.\nNew location = ..." in it, and a lot of configuration info is gone.

I changed the script so that when the correct files are copied to the ramdisk, another copy will be created in a backup directory. On next boot, the script will detect that the config file(s) is(are) corrupt and copy a backup version instead. The only problem is: while booting, the empty config file is used though. So after a reboot caused by power loss one must reboot again a second time correctly (via shell command), after this step, the backuped files a recovered in the /boot partition and the ramdisk as well.

TODO: change the boot process, so there already will the loss of config files be detected and the backup files used instead.