Create dietpi-recover
A script to recover erroneously deleted config files by the dietpi-ramdisk in conjunction with hard resets (power loss) of the Raspberry PI (in which case the current config files may not be re-written in the boot partition). The script should be called in the first line of the dietpi boot process to be sure, all depending scripts may use the recovered config files.