Update to blynserver install script
By not providing a config file to blynkserver, we get default configurations, which do not allow clients to access the admin page of blynkserver (at port 7443).
I edited the script so that it grabs a detailed configuration file from blynk's repo, edit the data folder path, and use the server config file as an argument to the service (instead of the data folder).

I only left the TO DO, so one can check one line of code there. I'm not sure if those are the best practices.

I also updated the blynk download link to the latest version.