Add "status" argument to dietpi-services
Very often I find useful to know which of my distpi services are actually running. systemctl provides this functionalities, but they are way too verbosed, and also show many other services along. I decided to implement the very simple patch to serve to with the propose of showing the status of dietpi services.

There may be improvements to make, but it works fine for me here.