v156
DietPi-Drive_Manager | We have reviewed "Mystic Meg's" claims that it is
possible to psychically unmount drives. Our team has personally tested
this claim, and, we can confirm this is simply not the case. Either
that, or our "Chinatown" crystal ball is not fully functional.
Therefore, we have deemed it necessary that users must "physically"
remove drives, instead of attempting to achieve it with their minds
alone. This also applies to keyboards.