Update Komga new setting for v0.107.0
I remove old configuration keys.
There is a configuration key in v0.107.0. 
file-hashing, a boolean indicating if Komga should compute a filehash for your files. This is required for the trash bin functionality to work, but can consume lots of resources on large libraries or slow hardware. Defaults to true.
I think set false is better for slow hardware. If the user want to use trash bin, they can change back to true later.