v8.15
Strange error on debian-ports Sid
---
[ INFO ] DietPi-Installer | APT install usr-is-merged, please wait...
Reading package lists...
Building dependency tree...
usrmerge is already the newest version (35).
0 upgraded, 0 newly installed, 0 to remove and 0 not upgraded.
---
as if the usr-is-merged package would be a symlink to the usrmerge one or so. It is not, checking package files and lists looks all correct. Let's with to install it via dpkg...