make the fat partition detection work on gpt
We were checking for the mbr type of 'c'. Instead of checking for that
and the GUID for a generic MS data partition, make things easier and
just check that it's a vfat filesystem. With that, the 4MB, and the
subsequent check for the label this appears to be enough belts and
suspenders.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>