dietpi-globals: G_CONFIG_INJECT: Fix detection of values with spaces
For whichever reason, G_CONFIG_INJECT assumed that the desired value is already set, even if the line contains more characters after the desired value. Probably the idea was to allow comments after the line. However, of course values themselves can contain spaces, and it can be intended to remove those parts.

This was found when trying to remove kernel command-line parameters from the end of the line, which did not work as of above bug.