dietpi-installer: drop tiny-initramfs and add initial ARM VM support
tiny-initramfs is not developed anymore since 9 years, even that Debian still keeps and maintains a package.

For systems which do not require an initramfs but where the kernel package has a dependency on "linux-initramfs-tool" (like RPi), it is the downside that its postinst trigger always generates an initramfs. I recently committed support for the INITRD variable, to skip the initramfs generation if a kernel package is installed/updated and sets INITRD=no. But there are various circumstances where other packages call it via trigger, which does "update-tirfs", and that command does not have an "-u" option like "update-initramfs", hence always generates initramfs images for all installed kernel versions, instead of renewing just existing ones. tiny-initramfs simply has no command to regenerate only existing initramfs images.

For VMs, the reason we used tiny-initramfs was that it builds tiny images with only the actually needed drivers, especially controller and filesystem drivers. When generating images via GitHub Actions, naturally needed controller drivers cannot be detected. When we supported only a small number of known VM virtualizers, it was no issue to add the small number of drivers for their common/default block device controllers, especially when we ship VM appliance images where we control this setting. But in the meantime we support a large number of virtualizers, with a growing number of drive controller options, and disk images without controller, where the drive controller can be selected on VM generation and default may have changed. Adding ARM VM support increases the possibilities further. Maintaining a list of controller drivers for all common options has become a burden and too large downside when our images do not support this or that VM setting, and changing it can prevent the image from booting.

We love small image sizes, but the burden and downside to keep using tiny-initramfs has become too large.