logind: enhance conditions for enabling/disabling it
QEMU Guest Agent requires it only on the version shipped with Bullseye. It is not needed for ACPI, but for the way QGA tries to shutdown the system as of a bug, resolved and backported down to the version shipped with Bookworm.

When generating VM images, pre-install dbus and unmask logind, instead of relying on this happening during first run setup. Have ACPI functionality up right from first boot.