dietpi-set_cpu: fix rare false min/max freq assignment
The script assumed the first left value of scaling_available_frequencies to be the lowest, and the last right value to be the highest available CPU frequency. If no min/max was explicitly set via dietpi-config or dietpi.txt, the first value was set as min and the last value as max. However, there are cases where the order is inverted. Adjust the awk call to loop through the values the find the actual lowest and highest regardless of order.