U-Boot: script and env updates
Add support for multiple device tree overlay prefixes. Especially in Rockhip vendor kernel, there are overlays with multiple prefixes available for the same SBC. Allow to assign multiple prefixes (seperated by space) to the overlay_prefix variable, and loop through all of them in the boot script. Check the existence of an overlay (and fixup) with each prefix, before loading it.

Apply known prefixes for Rockchip vendor kernel by default for new images.

Also obtain the (PART)UUID of the first partition of the device from where U-Boot was loaded, and add it as "partuuid" to the kernel command-line. Armbian uses this in its U-Boot flashing script. We did not do that, as all our images (which use this script) boot from a single partition. However, admins might move their root partition manually, and we aim to add support for it as well, e.g. minimal U-Boot-only or boot-only images, which contain the necessary parts only, and load the root filesystem from e.g. USB or NVMe drives, even if the SBC does not support it directly (no SPI bootloader storage). If this variable is not set in userland, we know an SPI bootloader was used.