dietpi-build: bootstrap with security repo added
but use "apt" base instead of "minbase", which excludes "required" packages, but includes "essential" and apt itself only.

There is currently no way to define multiple suites, like backports/updates/security. But when a stable release name is used as suite and no mirror is defined, mmdebstrap uses deb.debian.org and adds security and updates suites automatically, which is what we want.