dietpi-installer: readd dietpi-initramfs_cleanup
The initramfs-tools postrm cleanup script exists early on kernel upgrades, and performs the needed obsolete initrd.img removal on actual package removals only. This works well if the kernel package name chnges on every Linux version change, like on Debian, and RPi as well in the meantime, but with Armbian packaging, the same kernel package is upgraded, not replaced.

While without uInitrd, we'd not need to check for those anymore, but it does not hurt either.