ci: add workflow to update software URLs
It checks whether the expected latest version can be obtained for all architectures, and optionally checks for possibly newly supported architectures, riscv64 notably.

It replaces fallback URLs/versions with the newly found ones, in case, commits changes to the "update_urls" branch and opens a PR if it does not exist yet.