dietpi-set_software: add hint about Dropbear max 100 bytes password support
In dietpi.txt, we had a warning already, but it said "100 characters", while tests showed that the limit is at 100 bytes. Password managers tend to add a bunch of multi-byte characters and accents into the mix, which take 2 or more bytes each. Such characters are not problem, but 75 characters can than easily break the 100 bytes limit.