dietpi-software: GZDoom: link shared bzip2 and OpenAL libraries
With DYN_OPENAL=1 (default), the library is not linked, so the GZDoom starts without it, and no headers needed. It is detected and in case loaded "dynamically" (opposed to linking a dynamic/shared library).

This dynamic loading requires additional code. We add(ed) libopenal1 as package dependency anyway, so it makes sense to link it the classic way and skip the dynamic detection at GZDoom start.