NanoPi R4S/R5S/R6S: Move from netdev to native Ethernet driver LED triggers
We leave it as it is for R5C/R6C for now, which use one LED for WiFi, where whether/which trigger exists depends on the attached WiFi adapter.

For R2S, a native trigger exists for the WAN LED, but not for the LAN LED.

Generally, all LEDs on one SBC should behave the same.