dietpi-imager: regenerate ext4 journal after resize2fs and zerofree
So resize2fs might shrink the fs a bit more, and zerofree might not affect the journal. We will see whether re-enabling the journal can fail in this case, or whether it can be too small.

Point on that: resize2fs (on R/W mounted fs) does not raise the journal size. We might need to split dietpi-fs_partition_resize for a part to run before `systemd-remount-fs.service`, so that the journal size can be adjusted for the new filesystem size. This also generally allows the F2FS expansion. But there are still questions, like how to copy config files from trailing FAT config partition if the rootfs is still R/O mounted. Maybe the `mount -o remount,rw` and `mount -o remount,ro` are less problematic (filesystem being busy) before `systemd-remount-fs.service`, but with the flags and log file and disabling the service, ruling out a reboot loop, assuring that some debugging is possible etc, is non-trivial.