dietpi-software: X11: clear aDEPS variable
This is done within Download_Install() after packages are installed. But for X11, while we collect dependencies in this variable, we do not use Download_Install() anymore but install the dependencies manually. Hence the variable need to be cleared manually as well.

Usually, this has no negative impact. If another software installation calls Download_Install, it moves the download into background and calls APT again, which however is then just a noop. But if the DietPi-Dashboard is the next installed software option, and the nightly version is installed, the added URL check fails, since nighly.link returns a 404 on HEAD requests, even if the file actually exists and can be downloaded with GET.

Furthermore, clear the aDEPS variable instead of unsetting it (in Moonlight install block), to preserve its local state.

And in Kodi install block, no need to declare it local again, which is done at the top of the install function.