live patch 0: fix DietPi-Dashboard nightly installation
If unzip is not installed, or the aDEPS array is not empty, possible after X11 installation within same run, a HEAD request is done to check the nightly download URL, which always returns 404, even if the download via GET would succeed.

The patch assures that unzip is installed and the aDEPS array is empty before the download is attempted.