dietpi-imager: incrementally raise ext4 size until adding journal succeeds
Experimenting a bit more. Most builds succeed even with resize2fs minimizing without journal. But as expected, in some cases there is not sufficient space remaining for the journal. In this case, add 4 MiB per iteration as long as it still fails with the respective error.

Overall, it is probably not worth the hassle, but I am interested in the difference of image size and archive size this does have or has not.