dietpi-software: obtain latest pyenv patch version automatically
Keep the same subversion for now. microblog.pub definitely has some issues with Python 3.12. Home Assistant supports new versions very quickly, but some plugins may not. Maybe we can make it so that the lastest subversion with a patch version of "2" is used for Home Assistant, adding some time that way