dietpi-software: Moonlight (GUI): disable for Trixie on ARMv7
There is no Trixie suite yet, and the Bookworm packages depend on libssl3. On Trixie, libssl3t64 is provided instead, but it "provides" libssl3 for aarch64 only. The 64-bit time_t transition to prevent 2038 problems is relevant for 32-bit only. So for 64-bit, libssl3t64 equals libssl3, and the package name has been changed only to align with the 32-bit one.