dietpi-software: microblog.pub: update dependencies on install
Since the project is not currently maintained, for security reasons, and since 32-bit ARM builds are failing for some outdated dependency, update to latest dependencies, compatible based on pyproject.toml.

Also split the steps of Python, Poetry, dependencies and microblog.pub installation like we did not Home Assistant, so each command can be retried individually on failure, and it is more transparent which output belongs to which step.

Also, fix error "Error: The current project could not be installed: No file/folder found for package microblogpub" which happens due to package mode, used by default by Poetry. It requires the project directory to match the project name. However, we do not want to create microblog.pub packages but only manage dependencies for this particular instance with Poetry. Hence the package mode can be disabled.