ci: update_urls: disable phpBB update check for now
Cloudflare currently challenges all requests to phpbb.com, which can only be solved with a browser, since it is needed to press a butten, accept and send a cookie from that point on. This also breaks our phpBB installation option. Might be just temporary to fight some bot attack, phpBB has been contacted.