dietpi-patches: do not remove LED triggers for R5C/R6C
This was intended for the R5S/R6S only so far. The C models have a WiFi netdev trigger, and we want to keep it, as well as have it behaving consistently with the Ethernet LED triggers. The native driver triggers can not blink on activity, but only indicate a link. So either RX/TX activity indicator would need to be disabled for the WiFi LED, or netdev triggers for Ethernet LEDs need to be kept, which is what we intended to do so far.