dietpi-software: add function to select Python APT dependencies (#7625)
APT packages needed to compile certain Python modules depend on architecture, OS/Python version, and whether piwheels provides latest/needed pre-compiled wheels for ARMv6/7 or not. Currently, we maintain and regularly update these dependencies for each software title individually, even that it is basically the same in all cases, with some Python modules which needs to be compiled for multiple software titles: Pillow, NumPy, cryptography, ...

This function allows to merge the logic for all of them, with a single place to update dependencies if needed, and document the dependency trees more in detail without adding many lines on multiple places.