ci: vaultwarden: obtain compatible web vault version from vaultwarden
New web vault versions might not be compatible with vaultwarden releases yet, hence are also not marked as "latest" release, but kept pre-releases instead. Once a new vaultwarden version is released, respective web vault versions get the pre-release flag removed, and a "latest" tag is moved. But this does not always happen immediately.

Safest and always working solution is taking the web vault version from the Docker configuration file of the vaultwarden source.