general: replace policykit-1 with polkitd from Bookworm on
Additionally, do not install unused packages on GNUstep and do not purge unrelated ones on uninstall. GNUstep does not provide GUI poweroff/reboot functionality at all, but only two menu entries to exit and terminate the session respectively. This means that it does not make use of upower either. If anyone knows common extensions for this desktop environment which make use of polkit, or add said poweroff/reboot to the menu, let us know.

Add the migration as well to the Trixie upgrade script