dietpi-update: enhance update path for old Debian versions
Add recursive migration to compat branches for old Debian and DietPi versions. Currently, the updater switches to the next older compat branch only, and then performs the update, even if the Debian or DietPi version is too low for that compat branch as well. E.g. a Debian Stretch system on master branch would currently update to DietPi v8 on the "8" branch, while it should be migrated to the "stretch" branch.

With this change, the Check_Git_Migration function calls itself recursively, hence in above scenario, it migrates from "master" to "8" first, and from "8" to "stretch" branch next, as intended.

Furthermore, if the current Debian version is not supported anymore by the applied DietPi version, e.g. due to the previously missing recursive compat branch migration, G_DISTRO is set to 0, in which case the branch migration was skipped. This was an intentional change, to not switch branches if OS detection failed. However, it means that e.g. a Stretch system which unintentionally upgraded to DietPi v8 or above, will never be migated to any compat branch. The change is hence reverted, so that the compat branch migration is performed.

In scenarios, where the installed DietPi version is already too new for the Debian version, this however won't downgrade, just prevent further upgrades. We will solve any remaining such case individual, hence this change is mostly intended to prevent such issues in the future.