dietpi-software: TigerVNC: use new passwd path on Trixie
On local tests, the service migrated old to new path automatically, replacing /root/.vnc with a symlink to /root/.config/tigervnc. On CI test however this migration fails for unknown reasons: "tigervncserver: Could not migrate /root/.vnc to /root/.config/tigervnc."

However, it makes sense to use the correct path in the first place. But we hence know that no migration or patch is needed on our end via dietpi-update or Trixie upgrade script.