dietpi-obtain_hw_model: add support for older Debian versions and next testing Forky
We do not want to re-add support for old Debian versions, of course. This is only about defining the G_DISTRO variable, to allow correct recursive fallback compat branch migration of dietpi-update, if a system was unintentionally upgraded to an unsupported Debian version. This is was possible since dietpi-update did not try recursive branch migration previously. So if a Stretch system tried to upgrade to master branch, it was migrated to the next compat branch "8", and then the upgrade was applied, despite DietPi 8 requires Debian Buster or higher.

For recursive fallback to work, we need to keep setting G_DISTRO for down to the lowest Debian version we have a compat branch for, which is Debian Stretch with branch "stretch".