ci: dietpi-software test: skip Sonarr test on ARMv6
it fails with "If you're reading this, the MonoMod.RuntimeDetour selftest failed." while it succeeds on actual hardware. Not worth to debug the test only for this outdated version we ship for ARMv6, which we should better remove by times. Sonarr does not support ARMv6 since v4, since .NET does not. Pretty much the same for all .NET/Servarr apps, which are either not supported or shipped with pinned old versions.