dietpi-trixie-upgrade: fix potential pip conflicts
New versions of pip attempt to uninstall instances of modules installed via apt, and fail doing so. This is true as well for get-pip.py. If Python 3 was installed (as well) via dietpi-software, allow autoremoval of python3-pip/setuptools/wheel, the 3 modules installed by get-pip.py on the dietpi-software Python 3 reinstall. If there is a dependant, better let admins or our support review the individual cases.

Also purge the "mime-support" package, which is a transitional dummy package at least since Bullseye, and has been removed from the Trixie repo. For some reason it remains installed even without dependant and that it was priority "optional".