dietpi-software: WireGuard: enable unattended installs
dietpi-globals: G_WHIP: for menus and inputboxes, if G_WHIP_DEFAULT_ITEM and G_WHIP_NOCANCEL=1 are defined, a value is expected in any case. On interactive terminals, while no cancel button is shown, it is still possible to abort whiptail with ESC, or an error could lead to its failure. In this case, assign the default value as return value, so parent scripts can rely on a valid response as long as the default was valid.

An alternative would be to stay in the loop and show the dialogue again, but in case of a whiptail executable error, or falsely detected/defined non-interactive terminal, this could cause an endless loop or hang. Hence be convervative for now. We may go a step further after verifying the exact exit code of whiptail on ESC.

Solve sysctl write attempts differently: Create wrapper script which turns all calls into dry runs, which fail on wrong syntax and non-existing parameters, but does not fail if the container misses capabilities to change them.

For WireGuard, CAP_NET_ADMIN is granted to the container, so it can create the wg0 interface. This implies successful iptables calls. We might want to switch to a private network at some point, but then need to generate network-capable container images first, which is on the ToDo.