dietpi-software: MediaWiki: workaround bot blocking
For now just use some dummy user agent with curl. We will probably switch to GitHub API to obtain the version string, then forging the download URL from that.

Downloads of archives themselves is currently blocked as well as of today, but that should be fixed soon. To obtain the lastest version or URL, however, a better long-term solution is needed, which I am checkout out at the moment.