dietpi-config: always show RTC mode menu option
Until including Bullseye, hwclock was part of the pre-installed util-linux package. The menu option was only shown if the hwclock command succeeded, since otherwise fake-hwclock is required anyway, and there was no reason to show the option.

Since Bookworm however, the command has been moved into util-linux-extra, which is not pre-installed on DietPi. So far there is no strong reason to install util-linux-extra, but it makes sense to show the option even if the package is not installed (and hwclock hence doomed to fail). So if a real RTC is selected, util-linux-extra is installed, and afterwards the hwclock command is tested. If it fails, a dialog says so, and fake-hwclock remains installed.

Another stop could be to check for some RTC related dev/proc/sysfs node, do in case make this option more prominent, instead of keeping it in the adavanced options submenu only. fake-hwclock causes only minor issues if a real RTC is used, but a real RTC is useless if the hwclock command is not present.