fs_partition_resize: further fix for F2FS expansion
F2FS expansion cannot be done even for R/O mounted drives, neither does fsck.f2fs work for R/O mounts, which is a dedicated issue, causing `systemd-fsck-root.service` to always fail.

To hack around the limitation, create a loop device of the root drive, and perform resize on that one. Creating a loop device of the partition only causes a malformed checkpoint leading to an unrecoverable broken filesystem, at least when testing this on a fully booted R/O-mounted system. But creating the loop device for the whole root drive, scanning and expanding its partition, works.