fs_partition_resize: skip partition expansion if root drive not is missing
We obtain the drive device node path in a failsafe way based on known schemes. If that does not exist, it is a container system that is sandboxed enough so that informing the (host) kernel about partition changes is impossible anyway. Skip expansion gracefully, but print a message. Container images need to be expanded from the host in most cases.

Previously, we shipped those images without resize service enabled. Now this is needed for ext3/4 journal generation, and it does not hurt where supported. But in simple systemd-nspawn startups do not mount through any block device.

dietpi-installer: detect needed filesystem packages based on mounted filesystems instead of based on attached ones. This is mostly done since blkid requires the block device node to be mounted into containers, while findmnt works. However, it makes sense in general: the installer does not unmount any filesystem, and dietpi-drive_manager installs missing packages automatically on mount.