dietpi-update: exit script after reboot trigger
Generally it makes sense to exit scripts when triggering a reboot, else they continue to run until the init system kills it, which is usually not intended.

Furthermore, on first login, when started from dietpi-login, its STDOUT+STDERR are redirected to a log file on disk. If the rootfs is a network filesystem, this prevents it from being able to unmount, hanging the shutdown procedure.

Similarly, the login script itself should exit in these cases to avoid it looping to the next install stage if the init system takes long to initiate the shutdown procedure for some reason.