fs_partition_resize: reboot after adding ext4 journal
A reboot is strictly needed after adding a the filesytem journal, otherwise it is not practically used for this boot session. On reboot mount can then fail since the last metadata changes in the unused journal do not match those on disk. e2fsck fixes this, but in case only after a power cycle, if it was not triggered by another filesytem error, unclean shutdown, as part of regular fsck enforcement every 60 (default) mounts etc.