fs_partition_resize: journal creation requires R/W mount
(or an unmounted fs). Since this means that additional writes are done after journal creation and hence a metadata mismatch, an fsck is strictly required to fix this. /forcefsck is ignored by initramfs-tools, fsck.mode=force via cmdline is way too complex (the many different methods and not even assured to add cmdline entries, the need to remove it afterwards), mark filesystem as dirty.