dietpi-config: update hwclock option
The hwclock command is not actually needed, i.e. util-linux-extra does not need to be installed. Instead of hiding (or reverting) the option if hwclock fails, hide it if the RTC device nodes it would use do not exist. This is the typical reason if fails. Also add info that fake-hwclock does not hurt if an RTC is present, since it never turns back the clock.