dietpi-obtain_hw_model: never create missing /etc/.dietpi_hw_model_identifier
If the file is missing, it is most likely an RPi image with boot partition not mounted, e.g. if only the rootfs is mounted for chroot/container, of the boot mount failed for some reason. Do not hard code the generic device ID to /etc/.dietpi_hw_model_identifier in this case, so it is detected as RPi again once the boot mount is present.

This does not affect the generic device detection for the particular boot session: it will be ID 22 generic device regardless. But writing that ID to /etc/.dietpi_hw_model_identifier prevents the script from even reaching RPi detection code, until that file is manually removed.