dietpi-software: do not fail on sysctl for container boots
We currently skip apply sysctl settings in explicit container images, but we do not skip it when booting other images as containers, e.g. during image builds or for tests. We worked around it in our test script using "--dry-run". Use the same method inside DietPi code now if run as container without write access to the respective sysctl node.