dietpi-software: ProFTPD: config updates
Skip IdentLookups directives. The module is disabled by default since Bullseye (was was builtin , hence no need to disable it by directive that is supported only if the module has been explicitly enabled by the admin.

Furthermore, fix root user denial: Users listed in /etc/ftpusers non-intuitively are NOT allowed to use ProFTPD. We obviously thought it is the other way round, when commenting the "root" line. Skip that step and in case uncomment or re-add the line on dietpi-update. This however is not critical in terms of security, since the config contains "RootLogin off" anyway. But no reason to add the 2nd layer ASAP.