dietpi-software: Medusa: update systemd unit
The template is unnecessary bloated but misses some beneficial directives. We maintain an own one from now on. Also do not grant the "dietpi" group write access to internal files: this is needed only for accessing downloads and making generated media files accessible to players, of which both should be outside Medusa's internal file structure, config, and database.