timezone: skip unnecessary dpkg-reconfigure tzdata calls
When called non-interactively via `-f` option, all it does is reading the timezone from the /etc/localtime symlink and recreating that symlink with the obtained timezone. Hence it is a noop after we create that symlink within installer and on first boot.

Similarly, there is no point to remove /etc/localtime first.

/etc/timezone is deprecated and not actively created by the tzdata package since Bookworm, hence remove it in installer if present, like done before this commit.

Since we still support Bullseye until first release next year, do not remove /etc/timezone on first boot, but mimic what tzdata reconfigure does: update it if that file exists. Let's just follow Debian in this regards and stop treating it only after Debian decides to actively remove it (which was thought about and AFAIK temporarily the case in testing already).

There is no other place where