dietpi-banner: squash dashcolon and colon word-wrap options
For lines without a colon, probably no one would want to not have them indented at all. Hence squash the two options.

Rename "min" to "fixed", and only ask for the fixed indent column if that option was chosen. Assure that only numbers are accepted by the inputbox, it otherwise repeats with an error internally.

Remove the BW_MIN_USEABLE_SPACE variable handling, as it is currently unused.

Limit length of green lines to 53, matching the banner if word-wrap is disabled.

If word-wrapped lines are still too long, indent to fixed `INDENT_MIN = 3` independent of `INDENT_FIXED`.

Enhance `make_whitespace_offset` function: avoid printing leading spaces to variable, which can happen multiple times, but store only the number of spaces and return the actual spaces only once at the end.