dietpi-drive_manager: fix faulty unmount attempt on drive format
If a mounted partition is selected, but then full drive format mode, it was attempted to additionally unmount the partition after it has been unmounted and formatted in a previous "all partitions" loop. Do this only if there really is no partition table, but the filesystem was written on the bare drive without partition table.

Enhance format confirmation output a bit.