dietpi-trixie-upgrade: install nftables with dist-upgrade
Locally stored package lists are invalidated if the respective sources.list entries do not exist anymore, hence packages cannot be installed at this section of the code. Install it as part of the dist-upgrade step, which upgrades the fail2ban package as well. Use more transparent variable names for conditional arguments of upgrade and dist-upgrade calls.