dietpi-set_software: APT mirror: skip trailing slash for default Debian mirror
as we did in all other cases already: an alignment. The trailing slash is truncated internally anyway. We tend to add it for fully qualified URLs in other contexts to avoid a potential 301 redirect, but APT mirrors are never accessed like that, but only the (In)Release and Package(.*) files for the repo suite and component.